songart_project/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
lyricsgenius==3.0.1
numpy==1.24.2
Pillow==9.5.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
sumy==0.11.0
textblob==0.17.1
tqdm==4.65.0